feat(no-mixed-type): no-mixed-interface -> no-mixed-type
