--- docs/rules/no-mixed-interface.md
@@ -1,11 +1,0 @@
-# Restrict interfaces so that only members of the same kind of are allowed in them (no-mixed-interface)
-
-This rule enforces that an inteface only has one type of members, eg. only data properties or only functions.
-
-## Rule Details
-
-Mixing functions and data properties in the same interface is a sign of object-orientation style.
-
-## Options
-
-The rule does not accept any options.
